Unlocked blocks upon breaking if the breaker was the owner.
