+++ src/main/kotlin/com/dansplugins/factionsystem/listener/PlayerInteractListener.kt
@@ -209,62 +209,8 @@ class PlayerInteractListener(private val plugin: MedievalFactions) : Listener {
     }
 
     private fun unlock(player: Player, block: Block) {
-        val blockData = block.blockData
-        val holder = (block.state as? Chest)?.inventory?.holder
-        val blocks = if (blockData is Bisected) {
-            if (blockData.half == BOTTOM) {
-                listOf(block, block.getRelative(UP))
-            } else {
-                listOf(block, block.getRelative(DOWN))
-            }
-        } else if (holder is DoubleChest) {
-            val left = holder.leftSide as? Chest
-            val right = holder.rightSide as? Chest
-            listOfNotNull(left?.block, right?.block)
-        } else {
-            listOf(block)
-        }
-        plugin.server.scheduler.runTaskAsynchronously(plugin, Runnable {
-            val playerService = plugin.services.playerService
-            val mfPlayer = playerService.getPlayer(player) ?: playerService.save(MfPlayer(plugin, player)).onFailure {
-                player.sendMessage("$RED${plugin.language["BlockUnlockFailedToSavePlayer"]}")
-                plugin.logger.log(SEVERE, "Failed to save player: ${it.reason.message}", it.reason.cause)
-                return@Runnable
-            }
-            val lockService = plugin.services.lockService
-            val lockedBlocks = blocks.mapNotNull { lockService.getLockedBlock(MfBlockPosition.fromBukkitBlock(it)) }
-            val lockedBlock = lockedBlocks.firstOrNull()
-            if (lockedBlock == null) {
-                player.sendMessage("$RED${plugin.language["BlockUnlockNotLocked"]}")
-                return@Runnable
-            }
-            if (lockedBlock.playerId.value != mfPlayer.id.value) {
-                val lockOwner = playerService.getPlayer(lockedBlock.playerId)
-                val ownerName = if (lockOwner == null) {
-                    plugin.language["UnknownPlayer"]
-                } else {
-                    lockOwner.toBukkit().name ?: plugin.language["UnknownPlayer"]
-                }
-                if (!player.hasPermission("mf.force.unlock")) {
-                    player.sendMessage("$RED${plugin.language["BlockUnlockOwnedByOtherPlayer", ownerName]}")
-                    return@Runnable
-                } else {
-                    player.sendMessage("$RED${plugin.language["BlockUnlockProtectionBypassed", ownerName]}")
-                }
-            }
-            lockService.delete(lockedBlock.block).onFailure {
-                player.sendMessage("$RED${plugin.language["BlockUnlockFailedToDeleteBlock"]}")
-                plugin.logger.log(SEVERE, "Failed to delete block: ${it.reason.message}", it.reason.cause)
-                return@Runnable
-            }
-            val interactionService = plugin.services.interactionService
-            interactionService.setInteractionStatus(mfPlayer.id, null).onFailure {
-                player.sendMessage("$RED${plugin.language["BlockUnlockFailedToSetInteractionStatus"]}")
-                plugin.logger.log(SEVERE, "Failed to save interaction status: ${it.reason.message}", it.reason.cause)
-                return@Runnable
-            }
-            player.sendMessage("$RED${plugin.language["BlockUnlockSuccessful"]}")
-        })
+        val lockService = plugin.services.lockService
+        lockService.unlock(player, block)
     }
 
     private fun checkAccess(player: Player, block: Block) {
